+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Added support for applying Snowflake Cortex functions `Summarize` and `Sentiment`.
 - Added support for `Series.hist`.
 - Added support for list values in `Series.str.get`.
+- Added support for list values in `Series.str.slice`.
 
 #### Bug Fixes